+++ test/Misc/vla.cpp
@@ -327,17 +327,5 @@ TEST_CASE("Variable Length Array") {
 
     REQUIRE(std::string("world") == arr[0]);
   }
-
-  SECTION("JsonBuffer_strdup") {
-    int i = 16;
-    char vla[i];
-    strcpy(vla, "world");
-
-    DynamicJsonBuffer jsonBuffer;
-    const char* dup = jsonBuffer.strdup(vla);
-
-    REQUIRE(static_cast<const void*>(vla) != static_cast<const void*>(dup));
-    REQUIRE(std::string("world") == dup);
-  }
 }
 #endif